requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
altair==4.0.0
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
entrypoints==0.4
faiss-cpu==1.7.4
Flask==2.3.3
frozenlist==1.4.0
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.40
gpt==0.2
greenlet==3.0.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
langchain==0.0.278
langsmith==0.0.66
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.7
numpy==1.26.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.2
pandas==1.5.3
Pillow==10.1.0
protobuf==3.20.3
pyarrow==14.0.1
pydantic==1.10.13
pydeck==0.8.1b0
Pygments==2.17.1
Pympler==1.0.1
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rpds-py==0.13.1
semver==3.0.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.23
streamlit==1.18.1
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
Werkzeug==3.0.1
yarl==1.9.3
zipp==3.17.0
